creating all recipe tables

+++ database/migrations/2018_09_23_050826_create_ingredient_purchase_history_table.php
@@ -17,7 +17,7 @@ class CreateIngredientPurchaseHistoryTable extends Migration
             $table->increments('id');
             $table->integer('ingredient_id');
             $table->integer('store')->unsigned()->nullable();
-            $table->double('paid_price',6,2)->nullable();
+            $table->decimal('paid_price',6,2)->nullable();
             $table->string('brand')->nullable();
             $table->dateTime('purchase_ts')->nullable();
             $table->integer('product_rating')->unsigned()->nullable();
